added action_update_test.go
