+++ action_delete_test.go
@@ -48,6 +48,30 @@ func TestSysValidateDel(t *testing.T) {
 		err := a.SysValidation(h, ratingsDef, nil, []peer.ID{h.nodeID})
 		So(err, ShouldBeError)
 	})
+
+	Convey("validate del entry should fail if it doesn't match the del entry schema", t, func() {
+		err := sysValidateEntry(h, DelEntryDef, &GobEntry{C: ""}, nil)
+		So(err, ShouldNotBeNil)
+		So(err.Error(), ShouldEqual, "unexpected end of JSON input")
+
+		err = sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Fish":2}`}, nil)
+		So(err, ShouldNotBeNil)
+		So(err.Error(), ShouldEqual, "Validation Failed: validator %del failed: object property 'Hash' is required")
+
+		err = sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Hash": "not-a-hash"}`}, nil)
+		So(err, ShouldNotBeNil)
+		So(err.Error(), ShouldEqual, "Validation Failed: Error (input isn't valid multihash) when decoding Hash value 'not-a-hash'")
+
+		err = sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Hash": 1}`}, nil)
+		So(err, ShouldNotBeNil)
+		So(err.Error(), ShouldEqual, "Validation Failed: validator %del failed: object property 'Hash' validation failed: value is not a string (Kind: float64)")
+
+	})
+
+	Convey("validate del entry should succeed on valid entry", t, func() {
+		err := sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Hash": "QmUfY4WeqD3UUfczjdkoFQGEgCAVNf7rgFfjdeTbr7JF1C","Message": "obsolete"}`}, nil)
+		So(err, ShouldBeNil)
+	})
 }
 
 func TestSysDel(t *testing.T) {
