+++ action_test.go
@@ -184,77 +184,6 @@ func TestSysValidateEntry(t *testing.T) {
 		err := sysValidateEntry(h, HeadersEntryDef, &GobEntry{C: entryStr}, nil)
 		So(err, ShouldBeNil)
 	})
-
-	Convey("validate del entry should fail if it doesn't match the del entry schema", t, func() {
-		err := sysValidateEntry(h, DelEntryDef, &GobEntry{C: ""}, nil)
-		So(err, ShouldNotBeNil)
-		So(err.Error(), ShouldEqual, "unexpected end of JSON input")
-
-		err = sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Fish":2}`}, nil)
-		So(err, ShouldNotBeNil)
-		So(err.Error(), ShouldEqual, "Validation Failed: validator %del failed: object property 'Hash' is required")
-
-		err = sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Hash": "not-a-hash"}`}, nil)
-		So(err, ShouldNotBeNil)
-		So(err.Error(), ShouldEqual, "Validation Failed: Error (input isn't valid multihash) when decoding Hash value 'not-a-hash'")
-
-		err = sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Hash": 1}`}, nil)
-		So(err, ShouldNotBeNil)
-		So(err.Error(), ShouldEqual, "Validation Failed: validator %del failed: object property 'Hash' validation failed: value is not a string (Kind: float64)")
-
-	})
-
-	Convey("validate del entry should succeed on valid entry", t, func() {
-		err := sysValidateEntry(h, DelEntryDef, &GobEntry{C: `{"Hash": "QmUfY4WeqD3UUfczjdkoFQGEgCAVNf7rgFfjdeTbr7JF1C","Message": "obsolete"}`}, nil)
-		So(err, ShouldBeNil)
-	})
-
-}
-
-func TestSysValidateMod(t *testing.T) {
-	d, _, h := PrepareTestChain("test")
-	defer CleanupTestChain(h, d)
-
-	hash := commit(h, "evenNumbers", "2")
-	_, def, _ := h.GetEntryDef("evenNumbers")
-
-	/* This is actually bogus because it assumes we have the entry type in our chain but
-	           might be in a different chain.
-		Convey("it should check that entry types match on mod", t, func() {
-			a := NewModAction("oddNumbers", &GobEntry{}, hash)
-			err := a.SysValidation(h, def, nil, []peer.ID{h.nodeID})
-			So(err, ShouldEqual, ErrEntryTypeMismatch)
-		})
-	*/
-
-	Convey("it should check that entry isn't linking ", t, func() {
-		a := NewModAction("rating", &GobEntry{}, hash)
-		_, ratingsDef, _ := h.GetEntryDef("rating")
-		err := a.SysValidation(h, ratingsDef, nil, []peer.ID{h.nodeID})
-		So(err, ShouldEqual, ErrModInvalidForLinks)
-	})
-
-	Convey("it should check that entry validates", t, func() {
-		a := NewModAction("evenNumbers", nil, hash)
-		err := a.SysValidation(h, def, nil, []peer.ID{h.nodeID})
-		So(err, ShouldEqual, ErrNilEntryInvalid)
-	})
-
-	Convey("it should check that header isn't missing", t, func() {
-		a := NewModAction("evenNumbers", &GobEntry{}, hash)
-		err := a.SysValidation(h, def, nil, []peer.ID{h.nodeID})
-		So(err, ShouldBeError)
-		So(err, ShouldEqual, ErrModMissingHeader)
-	})
-
-	Convey("it should check that replaces is doesn't make a loop", t, func() {
-		a := NewModAction("evenNumbers", &GobEntry{}, hash)
-		a.header = &Header{EntryLink: hash}
-		err := a.SysValidation(h, def, nil, []peer.ID{h.nodeID})
-		So(err, ShouldBeError)
-		So(err, ShouldEqual, ErrModReplacesHashNotDifferent)
-	})
-
 }
 
 func TestCheckArgCount(t *testing.T) {
